Add libraries needed to compile in Makefile
Probably there are more libraries that what would be really needed.
The following step will be to remove them one by one and try to understand
what is needed and what is not needed. Additionally, I will likely want to
learn how to configure Emacs so that it doesn't complain about a missing include
(which is provided by the Makefile's flag), and try to understand how to troubleshoot
a missing library since I will likely have to deal with this problem in the future.