The code compiles with the command in the Makefile.
I should create another version starting from "scratch", adding things
as I go, so that I understand what each package does. I have made this
giant leap now just to make sure that I could actually arrive at the
"end result", but I should take things step by step, now, and get to
the bottom of this. Any error or mistakes now will be an opportunity
to learn, but I will have the knowledge that it is possible to suceed,
as I already have a clear "path to victory", so to speak.